Added Suri wool type and texture varient Added Rabbit Biome variations Added Pink eyes as a marker for Lutino rabbits coat variations to come Rabbits are shearable *shearing drops are still bugged

+++ src/main/java/mokiyoki/enhancedanimals/model/ModelEnhancedLlama.java
@@ -18,7 +18,7 @@ public class ModelEnhancedLlama extends ModelBase {
 
     private boolean banana = false;
     private boolean suri = false;
-    private float coatlength = 0.0F;
+    private int coatlength = 0;
 
     private final ModelRenderer head;
     private final ModelRenderer headShaved;
